+++ frontend/test/lit/test_gradient.py
@@ -18,7 +18,7 @@ import jax
 import numpy as np
 import pennylane as qml
 
-from catalyst import grad, qjit
+from catalyst import grad, jacobian, qjit
 
 
 # CHECK-LABEL: public @jit_grad_default
@@ -113,8 +113,8 @@ def grad_range_change():
         qml.RY(y, wires=1)
         return qml.expval(qml.PauliX(0)), qml.expval(qml.PauliY(1))
 
-    # CHECK: gradient.grad "fd" @f({{%[0-9]+}}, {{%[0-9]+}}) {diffArgIndices = dense<[0, 1]> : tensor<2xi64>, finiteDiffParam = 9.9999999999999995E-8 : f64} : (tensor<f64>, tensor<f64>) -> (tensor<f64>, tensor<f64>, tensor<f64>, tensor<f64>)
-    g = grad(f, argnum=[0, 1])
+    # CHECK: "gradient.grad"({{%[0-9]+}}, {{%[0-9]+}}) {callee = @f, diffArgIndices = dense<[0, 1]> : tensor<2xi64>, finiteDiffParam = 9.9999999999999995E-8 : f64, method = "fd"} : (tensor<f64>, tensor<f64>) -> (tensor<f64>, tensor<f64>, tensor<f64>, tensor<f64>)
+    g = jacobian(f, argnum=[0, 1])
     return g(jax.numpy.pi, jax.numpy.pi)