+++ frontend/test/pytest/test_measurements_results.py
@@ -207,7 +207,7 @@ class TestExpval:
 
         @qjit
         @qml.qnode(qml.device(backend, wires=3))
-        def expval6(x: float, y: float):
+        def expval(x: float, y: float):
             qml.RX(x, wires=0)
             qml.RY(y, wires=1)
             qml.RZ(0.1, wires=2)
@@ -218,11 +218,11 @@ class TestExpval:
             return qml.expval(qml.Hamiltonian(coeffs, obs))
 
         expected = np.array(-0.2715)
-        observed = expval6(np.pi / 4, np.pi / 3)
+        observed = expval(np.pi / 4, np.pi / 3)
         assert np.isclose(observed, expected)
 
         expected = np.array(-0.33695571)
-        observed = expval6(0.5, 0.8)
+        observed = expval(0.5, 0.8)
         assert np.isclose(observed, expected)
 
     def test_hamiltonian_2(self, backend):
@@ -230,7 +230,7 @@ class TestExpval:
 
         @qjit
         @qml.qnode(qml.device(backend, wires=2))
-        def expval6(x: float):
+        def expval(x: float):
             qml.RX(x, wires=0)
 
             coeff = np.array([0.8, 0.2])
@@ -247,11 +247,79 @@ class TestExpval:
             return qml.expval(qml.Hamiltonian(coeff, [obs, qml.PauliX(0)]))
 
         expected = np.array(0.2359798)
-        observed = expval6(np.pi / 4)
+        observed = expval(np.pi / 4)
         assert np.isclose(observed, expected)
 
         expected = np.array(-0.16)
-        observed = expval6(np.pi / 2)
+        observed = expval(np.pi / 2)
+        assert np.isclose(observed, expected)
+
+    def test_hamiltonian_3(self, backend):
+        """Test expval for nested Hamiltonian observable."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=2))
+        def expval(x: float):
+            qml.RX(x, wires=0)
+
+            coeff = np.array([0.8, 0.2])
+            obs_matrix = np.array(
+                [
+                    [0.5, 1.0j, 0.0, -3j],
+                    [-1.0j, -1.1, 0.0, -0.1],
+                    [0.0, 0.0, -0.9, 12.0],
+                    [3j, -0.1, 12.0, 0.0],
+                ]
+            )
+
+            obs = qml.Hermitian(obs_matrix, wires=[0, 1])
+            return qml.expval(
+                qml.Hamiltonian(
+                    coeff, [obs, qml.Hamiltonian([1, 1], [qml.PauliX(0), qml.PauliZ(1)])]
+                )
+            )
+
+        expected = np.array(0.4359798)
+        observed = expval(np.pi / 4)
+        assert np.isclose(observed, expected)
+
+        expected = np.array(0.04)
+        observed = expval(np.pi / 2)
+        assert np.isclose(observed, expected)
+
+    def test_hamiltonian_4(self, backend):
+        """Test expval with TensorObs and nested Hamiltonian observables."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def expval(x: float):
+            qml.RX(x, wires=0)
+            qml.RX(x + 1.0, wires=2)
+
+            coeff = np.array([0.8, 0.2])
+            obs_matrix = np.array(
+                [
+                    [0.5, 1.0j, 0.0, -3j],
+                    [-1.0j, -1.1, 0.0, -0.1],
+                    [0.0, 0.0, -0.9, 12.0],
+                    [3j, -0.1, 12.0, 0.0],
+                ]
+            )
+
+            obs = qml.Hermitian(obs_matrix, wires=[0, 1])
+            return qml.expval(
+                qml.Hamiltonian(
+                    coeff, [obs, qml.Hamiltonian([1, 1], [qml.PauliX(0), qml.PauliZ(1)])]
+                )
+                @ qml.PauliZ(2)
+            )
+
+        expected = np.array(-0.09284557)
+        observed = expval(np.pi / 4)
+        assert np.isclose(observed, expected)
+
+        expected = np.array(-0.03365884)
+        observed = expval(np.pi / 2)
         assert np.isclose(observed, expected)
 
 
@@ -484,6 +552,60 @@ class TestVar:
             expected,
         )
 
+    def test_hamiltonian_5(self, backend):
+        """Test variance with nested Hamiltonian observable."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x, y, coeffs):
+            qml.RX(x, wires=0)
+            qml.RY(y, wires=1)
+            qml.RZ(0.1, wires=2)
+
+            obs = [
+                qml.PauliX(0) @ qml.PauliZ(1),
+                qml.Hamiltonian(np.array([0.5]), [qml.PauliZ(0) @ qml.Hadamard(2)]),
+            ]
+
+            return qml.var(qml.Hamiltonian(coeffs, obs))
+
+        expected = np.array(0.1075)
+        coeffs = np.array([0.2, 0.6])
+        observed = circuit(np.pi / 4, np.pi / 3, coeffs)
+        assert np.isclose(
+            observed,
+            expected,
+        )
+
+    def test_hamiltonian_6(self, backend):
+        """Test variance with TensorObs and nested Hamiltonian observables."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x, y, coeffs):
+            qml.RX(x, wires=0)
+            qml.RY(y, wires=1)
+            qml.RZ(0.1, wires=2)
+
+            return qml.var(
+                qml.Hamiltonian(
+                    coeffs,
+                    [
+                        qml.PauliX(0)
+                        @ qml.PauliZ(1)
+                        @ qml.Hamiltonian(np.array([0.5]), [qml.Hadamard(2)])
+                    ],
+                )
+            )
+
+        expected = np.array(0.01)
+        coeffs = np.array([0.2])
+        observed = circuit(np.pi / 4, np.pi / 3, coeffs)
+        assert np.isclose(
+            observed,
+            expected,
+        )
+
 
 class TestOtherMeasurements:
     """Test other measurement processes."""
@@ -544,6 +666,323 @@ class TestOtherMeasurements:
 
         # qml.state
         assert np.allclose(result[5], expected(x, qml.state()))
+
+
+class TestNewArithmeticOps:
+    "Test PennyLane new arithmetic operators"
+
+    @pytest.mark.parametrize(
+        "meas, expected",
+        [[qml.expval, np.array(-0.70710678)], [qml.var, np.array(0.5)]],
+    )
+    def test_prod_xzi(self, meas, expected, backend):
+        """Test ``qml.ops.op_math.Prod`` converting to TensorObs."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x: float, y: float):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.RX(x + y, wires=2)
+            qml.CNOT(wires=[0, 1])
+            return meas(
+                qml.ops.op_math.Prod(
+                    qml.PauliX(wires=0), qml.PauliZ(wires=1), qml.Identity(wires=2)
+                )
+            )
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        assert np.allclose(expected, result)
+
+    @pytest.mark.parametrize(
+        "meas, expected",
+        [
+            [
+                qml.expval(
+                    qml.ops.op_math.Sum(
+                        qml.PauliX(wires=0), qml.PauliY(wires=1), qml.PauliZ(wires=2)
+                    )
+                ),
+                np.array(-1.41421356),
+            ],
+            [
+                qml.var(
+                    qml.ops.op_math.Sum(
+                        qml.PauliX(wires=0), qml.PauliY(wires=1), qml.PauliZ(wires=2)
+                    )
+                ),
+                np.array(2.0),
+            ],
+            [
+                qml.expval(qml.PauliX(wires=0) + qml.PauliY(wires=1) + qml.PauliZ(wires=2)),
+                np.array(-1.41421356),
+            ],
+            [
+                qml.var(qml.PauliX(wires=0) + qml.PauliY(wires=1) + qml.PauliZ(wires=2)),
+                np.array(2.0),
+            ],
+        ],
+    )
+    def test_sum_xyz(self, meas, expected, backend):
+        """Test ``qml.ops.op_math.Sum`` and ``+`` converting to HamiltonianObs.
+        with integer coefficients."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x: float, y: float):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.RX(x + y, wires=2)
+            qml.CNOT(wires=[0, 1])
+            return meas
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    @pytest.mark.parametrize(
+        "meas, expected",
+        [
+            [
+                qml.expval(
+                    qml.ops.op_math.Sum(
+                        qml.PauliX(wires=0),
+                        qml.PauliY(wires=1),
+                        qml.ops.op_math.SProd(0.5, qml.PauliZ(2)),
+                    )
+                ),
+                np.array(-1.06066017),
+            ],
+            [
+                qml.var(
+                    qml.ops.op_math.Sum(
+                        qml.ops.op_math.SProd(0.2, qml.PauliX(wires=0)),
+                        qml.ops.op_math.SProd(0.4, qml.PauliY(wires=1)),
+                        qml.ops.op_math.SProd(0.5, qml.PauliZ(wires=2)),
+                    )
+                ),
+                np.array(0.245),
+            ],
+            [
+                qml.expval(qml.PauliX(wires=0) + qml.PauliY(wires=1) + 0.5 * qml.PauliZ(wires=2)),
+                np.array(-1.06066017),
+            ],
+            [
+                qml.var(
+                    0.2 * qml.PauliX(wires=0)
+                    + 0.4 * qml.PauliY(wires=1)
+                    + 0.5 * qml.PauliZ(wires=2)
+                ),
+                np.array(0.245),
+            ],
+        ],
+    )
+    def test_sum_sprod_xyz(self, meas, expected, backend):
+        """Test ``qml.ops.op_math.Sum`` (``+``) and ``qml.ops.op_math.SProd`` (``*``)."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x: float, y: float):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.RX(x + y, wires=2)
+            qml.CNOT(wires=[0, 1])
+            return meas
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_mix_dunder(self, backend):
+        """Test ``*`` and ``@`` dunder methods."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=2))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.CNOT(wires=[0, 1])
+            return qml.expval(-0.5 * qml.PauliX(0) @ qml.PauliY(1))
+
+        result = circuit(np.pi / 4, np.pi / 4)
+        expected = np.array(0.25)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_sum_hermitian(self, backend):
+        """Test ``+`` with Hermitian observables."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=3))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.CNOT(wires=[0, 1])
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.expval(qml.Hermitian(A, wires=2) + qml.PauliX(0) + qml.Hermitian(A, wires=1))
+
+        result = circuit(np.pi / 4, np.pi / 4)
+        expected = np.array(2.0)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_prod_hermitian(self, backend):
+        """Test ``@`` with Hermitian observables."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=4))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.RY(x + y, wires=2)
+            qml.RY(x - y, wires=3)
+            qml.CNOT(wires=[0, 1])
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.expval(
+                qml.PauliZ(2)
+                @ qml.Hermitian(A, wires=1)
+                @ qml.PauliZ(0)
+                @ qml.Hermitian(A, wires=3)
+            )
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        expected = np.array(0.20710678)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_sprod_hermitian(self, backend):
+        """Test ``*`` and ``@`` with Hermitian observable."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=2))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.CNOT(wires=[0, 1])
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.expval(0.2 * qml.Hermitian(A, wires=1) @ qml.PauliZ(0))
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        expected = np.array(0.14142136)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_sum_sprod_prod_hermitian(self, backend):
+        """Test ``+`` of ``@`` with Hermitian observable."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=2))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+            qml.CNOT(wires=[0, 1])
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.expval(qml.Hermitian(A, wires=1) + qml.PauliZ(0) @ qml.PauliY(1))
+
+        result = circuit(np.pi / 4, np.pi)
+        expected = np.array(1.0)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_dunder_hermitian_1(self, backend):
+        """Test dunder methods with Hermitian observable to a HamiltonianObs."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=4))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+
+            qml.RX(x + y, wires=2)
+            qml.RX(y - x, wires=3)
+
+            qml.CNOT(wires=[0, 1])
+            qml.CNOT(wires=[0, 2])
+
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.expval(
+                0.2 * (qml.Hermitian(A, wires=1) + qml.PauliZ(0))
+                + 0.4 * (qml.PauliX(2) @ qml.PauliZ(3))
+            )
+
+        result = circuit(np.pi / 4, np.pi / 2)
+        expected = np.array(0.34142136)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
+
+    def test_dunder_hermitian_2(self, backend):
+        """Test dunder methods with Hermitian observable to a TensorObs."""
+
+        # Enabling new arithmetic operators
+        qml.operation.enable_new_opmath()
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=4))
+        def circuit(x, y):
+            qml.RX(x, wires=0)
+            qml.RX(y, wires=1)
+
+            qml.RX(x + y, wires=2)
+            qml.RX(y - x, wires=3)
+
+            qml.CNOT(wires=[0, 1])
+            qml.CNOT(wires=[0, 2])
+
+            A = np.array(
+                [[complex(1.0, 0.0), complex(2.0, 0.0)], [complex(2.0, 0.0), complex(1.0, 0.0)]]
+            )
+            return qml.var(
+                (qml.Hermitian(A, wires=1) + qml.PauliZ(0))
+                @ (0.5 * (qml.PauliX(2) @ qml.PauliZ(3)))
+            )
+
+        result = circuit(np.pi, np.pi)
+        expected = np.array(1.0)
+        assert np.allclose(expected, result)
+
+        qml.operation.disable_new_opmath()
 
 
 if __name__ == "__main__":