+++ doc/changelog.md
@@ -125,24 +125,35 @@
   array(112936.34906843)
   ```
 
-* Add support for Hamiltonian observables with integer coefficients.
-  [#248](https://github.com/PennyLaneAI/catalyst/pull/248)
+* Add support for the new PennyLane arithmetic operators.
+  [#250](https://github.com/PennyLaneAI/catalyst/pull/250)
 
-  For example, compiling the following circuit wasn't allowed before, but it is
-  now supported in Catalyst:
+  PennyLane is in the process of replacing ``Hamiltonian`` and ``Tensor`` observables with a set of
+  general arithmetic operators. These consist of
+  [Prod](https://docs.pennylane.ai/en/stable/code/api/pennylane.ops.op_math.Prod.html),
+  [Sum](https://docs.pennylane.ai/en/stable/code/api/pennylane.ops.op_math.Sum.html) and
+  [SProd](https://docs.pennylane.ai/en/stable/code/api/pennylane.ops.op_math.SProd.html).
+  By default, using dunder methods (eg. ``+``, ``-``, ``@``, ``*``) to combine operators with scalars or other
+  operators will create ``Hamiltonian``s and ``Tensor``s. However, these two methods will be deprecated in
+  coming releases of PennyLane.
+
+  To enable the new arithmetic operators, one can use ``Prod``, ``Sum``, and ``Sprod`` directly or activate them by
+  [enable_new_opmath](https://docs.pennylane.ai/en/stable/code/api/pennylane.operation.enable_new_opmath.html).
 
   ``` python
+  qml.operation.enable_new_opmath()
+  assert qml.operation.active_new_opmath()
+
   @qjit
   @qml.qnode(qml.device("lightning.qubit", wires=2))
   def circuit(x: float, y: float):
       qml.RX(x, wires=0)
-      qml.RY(y, wires=1)
+      qml.RX(y, wires=1)
+      qml.CNOT(wires=[0, 1])
+      return qml.expval(0.2 * qml.PauliX(wires=0) - 0.4 * qml.PauliY(wires=1))
 
-      coeffs = [1, 2]
-      obs = [qml.PauliZ(0), qml.PauliZ(1)]
-      return qml.expval(qml.Hamiltonian(coeffs, obs))
+  circuit(np.pi / 4, np.pi / 2)
   ```
-
 
 <h3>Improvements</h3>
 
@@ -199,6 +210,43 @@
   [#243](https://github.com/PennyLaneAI/catalyst/pull/243)
   [#247](https://github.com/PennyLaneAI/catalyst/pull/247)
 
+* Add support for Hamiltonian observables with integer coefficients.
+  [#248](https://github.com/PennyLaneAI/catalyst/pull/248)
+
+  For example, compiling the following circuit wasn't allowed before, but it is
+  now supported in Catalyst:
+
+  ``` python
+  @qjit
+  @qml.qnode(qml.device("lightning.qubit", wires=2))
+  def circuit(x: float, y: float):
+      qml.RX(x, wires=0)
+      qml.RY(y, wires=1)
+
+      coeffs = [1, 2]
+      obs = [qml.PauliZ(0), qml.PauliZ(1)]
+      return qml.expval(qml.Hamiltonian(coeffs, obs))
+  ```
+
+* Add support for nested Hamiltonian observables.
+  [#255](https://github.com/PennyLaneAI/catalyst/pull/255)
+
+  ``` python
+  @qjit
+  @qml.qnode(qml.device("lightning.qubit", wires=3))
+  def circuit(x, y, coeffs1, coeffs2):
+      qml.RX(x, wires=0)
+      qml.RX(y, wires=1)
+      qml.RY(x + y, wires=2)
+
+      obs = [
+          qml.PauliX(0) @ qml.PauliZ(1),
+          qml.Hamiltonian(coeffs1, [qml.PauliZ(0) @ qml.Hadamard(2)]),
+      ]
+
+      return qml.var(qml.Hamiltonian(coeffs2, obs))
+  ```
+
 <h3>Breaking changes</h3>
 
 * Support for Python 3.8 is dropped.
@@ -217,6 +265,11 @@
   will no longer work. This is more consistent with JAX's behaviour. However, if the variable
   ``list`` above is a JAX or Numpy array, the compilation still succeeds.
   [#221](https://github.com/PennyLaneAI/catalyst/pull/221)
+
+* The `catalyst.grad` function has been renamed to `catalyst.jacobian` and supports differentiation
+  of functions that return multiple or non-scalar outputs. A new `catalyst.grad` function has been
+  added that enforces that it is differentiating a function with a single scalar return value.
+  [#254](https://github.com/PennyLaneAI/catalyst/pull/254)
 
 <h3>Bug fixes</h3>